rag-app/requirements.txt
streamlit>=1.30.0
faiss-cpu>=1.7.4
ollama>=0.1.6
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.2
tiktoken>=0.5.2
numpy>=1.26.0
